dlaais

+++ dlaais/ants/registration_antspy.py
@@ -29,20 +29,18 @@ def rigid_reg_3d_nifti_groups(series_root, ref_file, in_filenames, out_filenames
     """
     if not os.path.exists(ref_file): return False
     fi = ants.image_read(ref_file)
-    fi_mask = ants.get_mask(fi, low_thresh=500, high_thresh=None)
+    fi_mask = ants.get_mask(fi, low_thresh=None, high_thresh=None)
     fi = ants.mask_image(fi, fi_mask)
-    ants.image_write(fi_mask, os.path.join(os.path.dirname(series_root), 'head_mask.nii.gz'))
-    ants.image_write(fi, os.path.join(os.path.dirname(series_root), 'CTP_P0.nii.gz'))
     for in_filename, out_filename in zip(in_filenames, out_filenames):
         print('working on %s - %s'%(series_root, in_filename))
         in_file = os.path.join(series_root, in_filename + '.nii.gz')
         out_file = os.path.join(series_root, out_filename)
         out_nii_file = out_file + '.nii.gz'
         if not os.path.exists(in_file): continue
-        #if os.path.exists(out_nii_file): continue
         mv = ants.image_read(in_file)
-        mv = ants.mask_image(mv, fi_mask)
-        mx = ants.registration(fixed=fi, moving=mv, type_of_transform='Rigid', mask=fi_mask,
+        mv_mask = ants.get_mask(mv, low_thresh=None, high_thresh=None)
+        mv = ants.mask_image(mv, mv_mask)
+        mx = ants.registration(fixed=fi, moving=mv, type_of_transform='QuickRigid', mask=fi_mask,
                                reg_iterations=(5, 5, 0), aff_iterations=(20, 10, 10, 10))
         mv_rigid = mx['warpedmovout']
         ants.image_write(mv_rigid, out_nii_file)
@@ -50,18 +48,19 @@ def rigid_reg_3d_nifti_groups(series_root, ref_file, in_filenames, out_filenames
 
 #--------------------------------------------------------------------------------------------------------
 #
-def rigid_reg_4d_nifti(patient_niix_root, dyn_1min_4d_niifilename, ref_nii_file='First'):
+def rigid_reg_4d_nifti(series_root, dyn4d_nii_filename, ref_nii_file='First', ext='_mc'):
     """
-    :param patient_niix_root:
-    :param dyn_1min_4d_niifile:
+    :param series_root:
+    :param dyn4d_nii_filename:
     :param ref_nii_file:        'First', 'Last', 'real file link'
+    :param ext:                 '_mc'
     :return:
     """
-    # decompose 4d nifti file
-    dyn_4d_niifile = os.path.join(patient_niix_root, dyn_1min_4d_niifilename + '.nii.gz')
-    dyn_4d_mc_niifile = os.path.join(patient_niix_root, dyn_1min_4d_niifilename + '_mc.nii.gz')
-    #if os.path.exists(dyn_4d_mc_niifile): return True
-    dyn_4d_mc_nii_root = os.path.join(patient_niix_root, dyn_1min_4d_niifilename)
+    # decompose 4d nifti file for quick rigid registration
+    dyn_4d_niifile = os.path.join(series_root, dyn4d_nii_filename + '.nii.gz')
+    dyn_4d_mc_niifile = os.path.join(series_root, dyn4d_nii_filename + ext + '.nii.gz')
+    if os.path.exists(dyn_4d_mc_niifile): return True
+    dyn_4d_mc_nii_root = os.path.join(series_root, dyn4d_nii_filename)
     if os.path.exists(dyn_4d_mc_nii_root): shutil.rmtree(dyn_4d_mc_nii_root)
     if not os.path.exists(dyn_4d_mc_nii_root):
         os.makedirs(dyn_4d_mc_nii_root, exist_ok=True)
@@ -75,9 +74,9 @@ def rigid_reg_4d_nifti(patient_niix_root, dyn_1min_4d_niifilename, ref_nii_file=
         for i in range(N):
             np_3d_img_1min = np_4d_img[:, :, :, i]
             nib_3d_img_1min = nib.Nifti1Image(np_3d_img_1min, affine=affine)
-            nib.save(nib_3d_img_1min, os.path.join(dyn_4d_mc_nii_root, dyn_1min_4d_niifilename + '_{:04d}.nii.gz'.format(i)))
-            in_filenames.append(dyn_1min_4d_niifilename + '_{:04d}'.format(i))
-            out_filenames.append(dyn_1min_4d_niifilename + '_{:04d}_mc'.format(i))
+            nib.save(nib_3d_img_1min, os.path.join(dyn_4d_mc_nii_root, dyn4d_nii_filename + '_{:04d}.nii.gz'.format(i)))
+            in_filenames.append(series_root + '_{:04d}'.format(i))
+            out_filenames.append(series_root + '_{:04d}'.format(i)+ext)
         np_4d_img = None
         nib_4d_img = None
 
@@ -107,14 +106,14 @@ def rigid_reg_4d_nifti(patient_niix_root, dyn_1min_4d_niifilename, ref_nii_file=
 
 #--------------------------------------------------------------------------------------------------------
 #
-def apply_mni_tx_4d_mc_nifti(patient_niix_root, dyn_1min_4d_niifilename, mni_txlist, ref_img):
+def apply_mni_tx_4d_mc_nifti(series_root, dyn4d_nii_filename, mni_txlist, ref_img, ext='_mni_nlin_sysm_09c_1mm'):
     """
     """
     # decompose 4d nifti file
-    dyn_4d_niifile = os.path.join(patient_niix_root, dyn_1min_4d_niifilename + '.nii.gz')
-    dyn_4d_mni_niifile = os.path.join(patient_niix_root, dyn_1min_4d_niifilename + '_mni.nii.gz')
+    dyn_4d_niifile = os.path.join(series_root, dyn4d_nii_filename + '.nii.gz')
+    dyn_4d_mni_niifile = os.path.join(series_root, dyn4d_nii_filename + ext + '.nii.gz')
     if os.path.exists(dyn_4d_mni_niifile): return True
-    dyn_4d_mni_nii_root = os.path.join(patient_niix_root, dyn_1min_4d_niifilename)
+    dyn_4d_mni_nii_root = os.path.join(series_root, dyn4d_nii_filename)
     if os.path.exists(dyn_4d_mni_nii_root): shutil.rmtree(dyn_4d_mni_nii_root)
     if not os.path.exists(dyn_4d_mni_nii_root):
         os.makedirs(dyn_4d_mni_nii_root, exist_ok=True)
@@ -128,9 +127,9 @@ def apply_mni_tx_4d_mc_nifti(patient_niix_root, dyn_1min_4d_niifilename, mni_txl
         for i in range(N):
             np_3d_img_1min = np_4d_img[:, :, :, i]
             nib_3d_img_1min = nib.Nifti1Image(np_3d_img_1min, affine=affine)
-            nib.save(nib_3d_img_1min, os.path.join(dyn_4d_mni_nii_root, dyn_1min_4d_niifilename + '_{:04d}.nii.gz'.format(i)))
-            in_filenames.append(dyn_1min_4d_niifilename + '_{:04d}'.format(i))
-            out_filenames.append(dyn_1min_4d_niifilename + '_{:04d}_mni'.format(i))
+            nib.save(nib_3d_img_1min, os.path.join(dyn_4d_mni_nii_root, dyn4d_nii_filename + '_{:04d}.nii.gz'.format(i)))
+            in_filenames.append(dyn4d_nii_filename + '_{:04d}'.format(i))
+            out_filenames.append(dyn4d_nii_filename + '_{:04d}'.format(i) + ext)
         np_4d_img = None
         nib_4d_img = None
 
@@ -151,3 +150,29 @@ def apply_mni_tx_4d_mc_nifti(patient_niix_root, dyn_1min_4d_niifilename, mni_txl
     # remove temp files
     if os.path.exists(dyn_4d_mni_nii_root): shutil.rmtree(dyn_4d_mni_nii_root)
     return True
+
+#--------------------------------------------------------------------------------------------------------
+#
+def deform_reg_3d_nifti_groups(series_root, ref_file, in_filenames, out_filenames):
+    """
+    :param series_root:
+    :param ref_file:
+    :param in_filenames:
+    :param out_filenames:
+    :return:
+    """
+    if not os.path.exists(ref_file): return False
+    fi = ants.image_read(ref_file)
+    fi_mask = ants.get_mask(fi, low_thresh=None, high_thresh=None)
+    fi = ants.mask_image(fi, fi_mask)
+    for in_filename, out_filename in zip(in_filenames, out_filenames):
+        print('working on %s - %s'%(series_root, in_filename))
+        in_file = os.path.join(series_root, in_filename + '.nii.gz')
+        out_file = os.path.join(series_root, out_filename)
+        out_nii_file = out_file + '.nii.gz'
+        if not os.path.exists(in_file): continue
+        mv = ants.image_read(in_file)
+        mx = ants.registration(fixed=fi, moving=mv, type_of_transform='QuickSyN')
+        mv_rigid = mx['warpedmovout']
+        ants.image_write(mv_rigid, out_nii_file)
+    return
